Fix bug with firstBillDay not marked as invalid in Configuration, after change.
Add logging to cache operations on UsageData and Configuration.

Fix bug with caching of prediction.  That's a function of time, which 
can't be cached.  This really, finally fixes Bug#3.

Move onPause() of superclass in Configuration, so that it fires before we
have invalidated the caches.  (Hmm, this may be bogus.)

Allow UsageData getPredictionAtBillMinutes to raise an out-of-bounds
exception if there's nothing to predict.  Returning zero was bogus!

Move prediction into own method, outside catching log info.  It can 
be cached.

Don't display prediction if it's impossible/bogus.